some lints
